requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
simple-websocket==1.0.0
cloudinary==1.36.0
bleach==6.1.0
dnspython==2.4.2

